+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.7.0] - 2021-08-11
+## Added
+- `queryByPage` for customized `forEachPage` queries
+- `storeSymbols.pageBreak` Symbol to allow `forEachPage` and `queryByPage` to be halted early
 
 ## [1.6.0] - 2021-08-03
 ## Added